Actualización5.5

+++ catalogue/views.py
@@ -1,10 +1,10 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect, get_object_or_404
 from catalogue.forms import ProductoForm, ServicioForm
 from catalogue.models import Producto, Categoria, Servicio
 
 
 # Create your views here.
-def productos(request):
+def producto(request):
     return render(request, 'templateCatalogue/productos.html')
 
 #Productos
@@ -14,38 +14,50 @@ def crear_producto(request):
         form = ProductoForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('mostrar_productos')
+            return redirect('mostrarProducto') # <-- SOLUCIÓN: sin 's'
     else:
         form = ProductoForm()
     return render(request, 'templateCatalogue/crear_producto.html', {'form': form})# no olvidar crear la plantilla crear_producto.html
 
 
-def mostrar_productos(request):
-    productos = Producto.objects.all()
+def mostrar_producto(request):
+    producto = Producto.objects.all()
     categoria = Categoria.objects.all()
 
     data={
-        'productos':productos,
+        'producto':producto,
         'categoria':categoria
         
     }   
     return render(request, 'templateCatalogue/productos.html',data)
 
-def editar_producto(request, id):
-    producto = Producto.objects.get(id=id)
+def cargar_producto(request,id):
+    producto = get_object_or_404(Producto,id=id)
+    form = ProductoForm(instance=producto)
+    return render(request,'templateCatalogue/productoEdit.html',{'form':form,'producto':producto})
+
+
+def modificar_producto(request,id):
+    producto = get_object_or_404(Producto,id=id)
     if request.method == 'POST':
-        form = ProductoForm(request.POST, instance=producto)
+        form = ProductoForm(request.POST,instance=producto)
         if form.is_valid():
-            form.save()
-            return redirect('mostrar_productos')
+            form.save()   
+        return redirect('mostrarProducto')
     else:
         form = ProductoForm(instance=producto)
-    return render(request, 'templateCatalogue/editar_producto.html', {'form': form})
+    return render(request, 'templateCatalogue/productoEdit.html', {'form': form,})
 
-def eliminar_producto(request, id):
-    producto = Producto.objects.get(id=id)
-    producto.delete()
-    return redirect('mostrar_productos')
+
+def eliminar_producto(request,id):
+    producto = get_object_or_404(Producto,id=id)
+
+    if request.method == 'POST':
+        producto.delete()
+        return redirect('mostrarProducto')
+    return render(request,'templateCatalogue/eliminarProducto.html',{'producto':producto})
+
+
 
 
 #Categorias
@@ -71,4 +83,3 @@ def crear_servicio(request):
     else:
         form = ServicioForm()
     return render(request, 'templateCatalogue/servicioAdd.html', {'form': form})
-
